// frontend/src/app/model/member/additional-information.ts
export interface AdditionalInformation {
  idAdditional?: number;
  repeatYear?: boolean;
  studyingAnotherDirection?: boolean;
  nameUniversity?: string;
  department?: string;
  fieldOfStudy?: string;
  profile?: string;
  yearStartStudy?: string;
  workPlacementAndYear?: string;
  studiedAnotherUniversity?: boolean;
  nameAnotherUniversity?: string;
  departmentAnother?: string;
  fieldOfStudyAnother?: string;
  profileAnother?: string;
  yearStartAndEndStudyAnother?: string;
  workPlacementAndYearAnother?: string;
}